+++ internal/provisioner/helm/controllers/bundle_controller.go
@@ -35,8 +35,8 @@ import (
 	helm "github.com/operator-framework/rukpak/internal/provisioner/helm/types"
 	"github.com/operator-framework/rukpak/internal/source"
 	"github.com/operator-framework/rukpak/internal/storage"
-	updater "github.com/operator-framework/rukpak/internal/updater/bundle"
 	"github.com/operator-framework/rukpak/internal/util"
+	updater "github.com/operator-framework/rukpak/pkg/updater/bundle"
 )
 
 // BundleReconciler reconciles a Bundle object
