+++ internal/provisioner/helm/controllers/bundledeployment_controller.go
@@ -45,8 +45,8 @@ import (
 	rukpakv1alpha1 "github.com/operator-framework/rukpak/api/v1alpha1"
 	helm "github.com/operator-framework/rukpak/internal/provisioner/helm/types"
 	"github.com/operator-framework/rukpak/internal/storage"
-	updater "github.com/operator-framework/rukpak/internal/updater/bundle-deployment"
 	"github.com/operator-framework/rukpak/internal/util"
+	updater "github.com/operator-framework/rukpak/pkg/updater/bundle-deployment"
 )
 
 // BundleDeploymentReconciler reconciles a BundleDeployment object
