+++ pkg/updater/bundle-deployment/updater_test.go
@@ -14,7 +14,7 @@ import (
 
 	rukpakv1alpha1 "github.com/operator-framework/rukpak/api/v1alpha1"
 	plain "github.com/operator-framework/rukpak/internal/provisioner/plain/types"
-	bundledeployment "github.com/operator-framework/rukpak/internal/updater/bundle-deployment"
+	bundledeployment "github.com/operator-framework/rukpak/pkg/updater/bundle-deployment"
 )
 
 var _ = Describe("Updater", func() {
