+++ pkg/updater/bundle/updater_test.go
@@ -14,7 +14,7 @@ import (
 
 	rukpakv1alpha1 "github.com/operator-framework/rukpak/api/v1alpha1"
 	plain "github.com/operator-framework/rukpak/internal/provisioner/plain/types"
-	"github.com/operator-framework/rukpak/internal/updater/bundle"
+	"github.com/operator-framework/rukpak/pkg/updater/bundle"
 )
 
 var _ = Describe("Updater", func() {
